Add separate file for generateOTP & emailService
