+++ backend/config/generateOtp.js
@@ -1,0 +1,6 @@
+
+const generateOtp = () => {
+    return Math.floor(100000 + Math.random() * 900000).toString(); // 6-digit OTP
+};
+
+module.exports = generateOtp;
